requirements.txt
streamlit==1.30.0
qdrant-client==1.7.3
sentence-transformers==2.2.2
transformers==4.32.1
huggingface-hub==0.16.4
pillow==10.1.0
numpy==1.26.2
pandas==2.1.2
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
